Remove Keywords, reorganization
We can probably do without keywords to keep a bit of flexibility in
Braai. The intention is that the first command encountered will be the
name of the retriever that we're going to invoke, the path being the
following dots and commands that occur after it. This path can be some
signal that the retriever looks at to know which thing it should
retrieve.

Also roughly reorganized the lexer to have public API first, helper
functions, then stateFns so it's a bit easier to navigate