Improve Coverage from Real Braai Tags
There were two previously uncovered cases. The first was Braai tags that
had a single closing bracket: {{ photo_gallery 'Foo' }, for example.
This caused the compiler to panic. The second was having real file names
in a bracketed argument... the characters '-' and '.' were not
acceptable, so the string was rejected by the compiler. Cases for these
have been added along with the appropriate error cases