Provide a way to define handlers
You can now define handlers by creating a new HandlerMux. This can take
a HandlerFunc, which receives the BraaiTagNode as an argument and should
return a final string and an error. Additionally, there is a Handle
method which uses reflection to call methods on a struct as it
encounters them, erroring appropriately if an unknown method is
encountered in the Braai source.